server/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn==0.24.0

# WebSocket Support
websockets==12.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# AI Integration
openai==1.3.7

# HTTP Client
httpx==0.25.2
requests==2.31.0

# Environment & Configuration
python-dotenv==1.0.0

# CORS
fastapi-cors==0.0.6

# Logging
loguru==0.7.2

# Data Validation
pydantic==2.5.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Development
black==23.11.0
flake8==6.1.0
